tambah custom_tags.py, edit sesuai grup

+++ kajian/ajax_datatable.py
@@ -1,6 +1,6 @@
 from ajax_datatable.views import AjaxDatatableView
 from django.core.exceptions import PermissionDenied
-from django.utils.translation import gettext_lazy as _
+from django.contrib.auth.models import User, Group
 
 from .models import Kajian
 
@@ -26,8 +26,9 @@ class KajianAjaxView(AjaxDatatableView):
     ]
 
     def customize_row(self, row, obj):
+        grup = User.objects.values("groups").get(username=self.request.user)
+        print(grup)
         file = self.model.objects.get(id=row["pk"])
-        print(row["pk"])
         if file.file:
             path_file = file.file.url
         else:
@@ -35,25 +36,39 @@ class KajianAjaxView(AjaxDatatableView):
         row['file'] = """
         <a href="%s">file</a>
         """ % path_file
-        row['action'] = f"""
-                    <a href="#" class="btn btn-primary" id="add" 
-                    onclick="add('{row['pk']}'); " >
-                       Tambah Progress
-                    </a>
-                    <a href="#" class="btn btn-info btn-edit" id="edit"
-                    onclick="edit('{row['pk']}'); " >
-                       Edit
-                    </a>
-                    <a href="/kajian_delete/{row['pk']}" class="btn btn-danger" data-toggle="modal"
-                    data-target="#delete-item-modal"
-                    id="delete-item"
-                    >
-                    Delete
-                    </a>
-                """
+        if grup["groups"] == 3:
+            row['action'] = f"""
+                        <a href="#" class="btn btn-primary" id="add" 
+                        onclick="add('{row['pk']}'); " >
+                           Tambah Progress
+                        </a>
+                        <a href="#" class="btn btn-info btn-edit" id="edit"
+                        onclick="edit('{row['pk']}'); " >
+                           Edit
+                        </a>
+                        <a href="/kajian_delete/{row['pk']}" class="btn btn-danger" data-toggle="modal"
+                        data-target="#delete-item-modal"
+                        id="delete-item"
+                        >
+                        Delete
+                        </a>
+                    """
+        elif grup["groups"] == 1:
+            row['action'] = f"""
+                                    <a href="#" class="btn btn-primary" id="add" 
+                                    onclick="add('{row['pk']}'); " >
+                                       Detail
+                                    </a>
+                                    
+                                """
 
     def get_initial_queryset(self, request=None):
+        grup = User.objects.values("groups").get(username=self.request.user)
         if not request.user.is_authenticated:
             raise PermissionDenied
-        qs = self.model.objects.filter(created_by=self.request.user)
-        return qs
+        elif grup["groups"] == 3:
+            return self.model.objects.filter(created_by=self.request.user)
+        elif grup["groups"] == 1:
+            return self.model.objects.all()
+        else:
+            return self.model.objects.all()
